# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.AdminSearchUserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      banned_at:
        example: "2024-02-01T10:30:00Z"
        type: string
      bio:
        example: Movie enthusiast
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      role:
        example: user
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
  handlers.ChangePasswordRequest:
    properties:
      current_password:
        example: oldpassword123
        type: string
      new_password:
        example: newpassword456
        maxLength: 72
        minLength: 8
        type: string
    required:
    - new_password
    type: object
...
  handlers.PublicUserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      banned_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      bio:
        example: Movie enthusiast
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_followed_by:
        example: false
        type: boolean
      is_following:
        example: true
        type: boolean
      role:
        example: user
        type: string
      stats:
        $ref: '#/definitions/handlers.UserStats'
      username:
        example: johndoe
        type: string
      website:
        example: https://example.com
        type: string
    type: object
...
  handlers.UpdatePreferencesRequest:
    properties:
      locale:
        enum:
        - en
        - fr
        - es
        example: fr
        type: string
      theme:
        enum:
        - light
        - dark
        - system
        example: dark
        type: string
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      bio:
        example: Updated bio
        maxLength: 500
        type: string
      preferences:
        $ref: '#/definitions/handlers.UpdatePreferencesRequest'
      username:
        example: newusername
        maxLength: 50
        minLength: 3
        type: string
      website:
        example: https://newwebsite.com
        type: string
    type: object
  handlers.UserPreferences:
    properties:
      locale:
        example: en
        type: string
      theme:
        example: system
        type: string
    type: object
  handlers.UserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      banned_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      bio:
        example: Movie enthusiast
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      preferences:
        $ref: '#/definitions/handlers.UserPreferences'
      role:
        example: user
        type: string
      stats:
        $ref: '#/definitions/handlers.UserStats'
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      username:
        example: johndoe
        type: string
      website:
        example: https://example.com
        type: string
    type: object
  handlers.UserStats:
    properties:
      followers_count:
        example: 150
        type: integer
      following_count:
        example: 75
        type: integer
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/response.Pagination'
      success:
        type: boolean
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@duskforge.io
    name: API Support
    url: https://github.com/duskforge/api
  description: A movie collection and review API powered by TMDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Duskforge API
  version: "1.0"
paths:
...
  /users/{userId}:
    get:
      consumes:
      - application/json
      description: Get the public profile of a user by their ID. If authenticated,
        includes follow relationship info. Follower and following counts always exclude
        banned users. Returns 404 if the user is banned (non-admin callers). Returns
        403 if there is a block between the authenticated user and the target user.
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User public profile
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.PublicUserResponse'
              type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found or banned
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
...
  /users/me:
    delete:
      consumes:
      - application/json
      description: Permanently delete the currently authenticated user's account.
        Super-admin accounts cannot be deleted.
      produces:
      - application/json
      responses:
        "204":
          description: Account deleted successfully
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Super-admin account cannot be deleted
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get the profile of the currently authenticated user. Follower and
        following counts always exclude banned users.
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update the profile of the currently authenticated user
      parameters:
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.UserResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Username already taken
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
...
  /users/me/avatar:
    delete:
      description: Remove the avatar image of the current user
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile without avatar
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete avatar
      tags:
      - users
    put:
      consumes:
      - multipart/form-data
      description: Upload a new avatar image for the current user. Replaces existing
        avatar if one exists. Accepts JPEG, PNG, GIF, and WebP (max 5MB).
      parameters:
      - description: Avatar image file
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile with new avatar
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.UserResponse'
              type: object
        "400":
          description: Invalid file or file too large
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Upload or update avatar
      tags:
      - users
...
  /users/me/password:
    put:
      consumes:
      - application/json
      description: Change the password of the currently authenticated user. OAuth-only
        users can omit current_password to set a password for the first time.
      parameters:
      - description: New password (current_password optional for OAuth-only accounts)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request body or password too short/long
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized or incorrect current password
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Change or set password
      tags:
      - users
  /users/search:
    get:
      consumes:
      - application/json
      description: Search and browse users by username with sorting and pagination.
        Regular users only see non-admin, non-banned accounts. Admins and super-admins
        see all accounts with additional details. If query is omitted, returns all
        visible users.
      parameters:
      - description: Search query (username)
        in: query
        name: query
        type: string
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - default: 20
        description: Number of items to return (max 100)
        in: query
        name: limit
        type: integer
      - description: Sort field with direction prefix (+asc, -desc)
        enum:
        - +username
        - -username
        - +created_at
        - -created_at
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results (admin view)
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.AdminSearchUserResponse'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"